docs: moving Styles quick start up
