+++ packages/documentation/src/Styles/Typography.stories.tsx
@@ -214,94 +214,6 @@ const CommonTemplate = ({
 	);
 };
 
-export const QuickStart: Story<any> = () => (
-	<div className="prose prose-dark dark:prose-lighter">
-		<h1>Typography</h1>
-		<p>
-			To use the custom typography styles provided by UI Components, you need to
-			use and configure Tailwind accordingly.
-		</p>
-		<h2>Installation</h2>
-		<p>
-			You have to install the <code>@versini/ui-styles</code> package:
-		</p>
-		<pre>
-			<code>$ npm install --save @versini/ui-styles</code>
-		</pre>
-
-		<h2>Configuration</h2>
-		<p>
-			You need to configure your application to use TailwindCSS and our
-			TailwindCSS plugin:
-		</p>
-		<pre>
-			<code>
-				{`// tailwind.config.js
-/** @type {import('tailwindcss').Config} */
-
-import { twPlugin } from "@versini/ui-styles";
-
-export default twPlugin.merge({
-  // this is an example, you can change the path to your files
-  content: ["./src/**/*.{js,ts,jsx,tsx}"],
-});`}
-			</code>
-		</pre>
-
-		<h2>Usage</h2>
-		<p>
-			Now you can use the <code>prose-dark</code>,<code>prose-light</code> and{" "}
-			<code>prose-lighter</code> classes to add our typography styles to any
-			vanilla HTML:
-		</p>
-		<pre>
-			<code>
-				{`<article class="prose prose-dark">
-  <h1>Garlic bread with cheese: What the science tells us</h1>
-  <p>
-    For years parents have espoused the health benefits of eating garlic bread with cheese to their
-    children, with the food earning such an iconic status in our culture that kids will often dress
-    up as warm, cheesy loaf for Halloween.
-  </p>
-  <p>
-    But a recent study shows that the celebrated appetizer may be linked to a series of rabies cases
-    springing up around the country.
-  </p>
-  <!-- ... -->
-</article>`}
-			</code>
-		</pre>
-
-		<p>
-			<strong>NOTE:</strong> always include the <code>prose</code> class when
-			adding a typography style (dark, light, or lighter).
-		</p>
-
-		<h2>Dark mode</h2>
-		<p>
-			Our typography styles are designed to work in dark mode, too. Just add the
-			<code>dark:</code> prefix to the classes:
-		</p>
-		<pre>
-			<code>
-				{`<article class="prose prose-dark dark:prose-light">
-  <h1>Garlic bread with cheese: What the science tells us</h1>
-  <p>
-    For years parents have espoused the health benefits of eating garlic bread with cheese to their
-    children, with the food earning such an iconic status in our culture that kids will often dress
-    up as warm, cheesy loaf for Halloween.
-  </p>
-  <p>
-    But a recent study shows that the celebrated appetizer may be linked to a series of rabies cases
-    springing up around the country.
-  </p>
-  <!-- ... -->
-</article>`}
-			</code>
-		</pre>
-	</div>
-);
-
 export const Light: Story<any> = () => (
 	<div className="bg-slate-900 p-11">
 		<div className="prose prose-light max-w-none">
